Add custom entrypoint to support direct execution of module.
We use npm to start services, but it does not relay system signals to
the process correctly in docker environment (see
https://github.com/goldbergyoni/nodebestpractices/blob/master/sections/docker/bootstrap-using-node.md
for technical details). This patch introduces entrypoint that process custom
run command syntax which runs module using node directly:
```
  run MODULE [options]

  MODULE    Module to run (e.g. app, downloader, uploader, etc.)
  options   Any valid node params that needs to be added (e.g. --max-old-space-size=4096).
```

If command is not mathing syntax, it will be passed to
`docker-entrypoint.sh` provided by node image.